AUTH feat(verify-email): refactor verification process into separate components and actions
