+++ containers/frontend/web/src/features/auth/verify-email/index.ts
@@ -1,0 +1,1 @@
+export { VerifyEmailFeature } from './verify-email';
